// src/API.ts
export type Note = {
  __typename: "Note",
  id: string,
  clientId?: string | null,
  name: string,
  description?: string | null,
  completed?: boolean | null,
  createdAt: string,
  updatedAt: string,
};
// ...
export type OnCreateNoteSubscription = {
  onCreateNote?:  {
    __typename: "Note",
    id: string,
    clientId?: string | null,
    name: string,
    description?: string | null,
    completed?: boolean | null,
    createdAt: string,
    updatedAt: string,
  } | null,
};

export type OnUpdateNoteSubscriptionVariables = {
  filter?: ModelSubscriptionNoteFilterInput | null,
};

export type OnUpdateNoteSubscription = {
  onUpdateNote?:  {
    __typename: "Note",
    id: string,
    clientId?: string | null,
    name: string,
    description?: string | null,
    completed?: boolean | null,
    createdAt: string,
    updatedAt: string,
  } | null,
};

export type OnDeleteNoteSubscriptionVariables = {
  filter?: ModelSubscriptionNoteFilterInput | null,
};

export type OnDeleteNoteSubscription = {
  onDeleteNote?:  {
    __typename: "Note",
    id: string,
    clientId?: string | null,
    name: string,
    description?: string | null,
    completed?: boolean | null,
    createdAt: string,
    updatedAt: string,
  } | null,
};